refactor: architecture cleanup sprint (#8)
- Fix string-based exception check for ApprovalRequiredError (C2)
- Move CronDefinition/CronOutputConfig to core/config/models (C1)
- Move AgentWorkspace to core/workspace (M1)
- Add public API on QueueManager and MessageProcessor (M2)
- Consolidate duplicate _resolve_user_name to core/utils (M3)
- DRY builtin config extraction in BuiltinLoader (M4)
- Decompose WorkspaceRunner.__init__ into init helpers (M6)
- Replace Any types in LifecycleManager with typed signatures (M7)
- Remove stale ruff per-file-ignores for nonexistent dirs (m1)
- Relocate ApprovalGateManager to runtime/approval (m4)
