+++ openpaw/core/config/models.py
@@ -7,6 +7,7 @@ For loading and merging logic, see loader.py.
 from pathlib import Path
 from typing import Any, Literal
 
+from apscheduler.triggers.cron import CronTrigger
 from pydantic import BaseModel, Field, field_validator
 
 
@@ -343,6 +344,48 @@ class LifecycleConfig(BaseModel):
     notify_startup: bool = Field(default=False, description="Send notification when workspace starts")
     notify_shutdown: bool = Field(default=True, description="Send notification when workspace stops")
     notify_auto_compact: bool = Field(default=True, description="Send notification on auto-compact")
+
+
+class CronOutputConfig(BaseModel):
+    """Output routing configuration for a cron job."""
+
+    channel: str = Field(description="Channel type (telegram, discord, etc.)")
+    chat_id: int | None = Field(default=None, description="Telegram chat ID")
+    guild_id: int | None = Field(default=None, description="Discord guild ID")
+    channel_id: int | None = Field(default=None, description="Discord channel ID")
+    delivery: str = Field(
+        default="channel",
+        description="Delivery mode: channel, agent, or both",
+    )
+
+    @field_validator("delivery")
+    @classmethod
+    def validate_delivery(cls, v: str) -> str:
+        """Validate delivery mode is one of the allowed values."""
+        allowed = {"channel", "agent", "both"}
+        if v not in allowed:
+            raise ValueError(f"Invalid delivery mode '{v}'. Must be one of: {allowed}")
+        return v
+
+
+class CronDefinition(BaseModel):
+    """Definition of a single cron job from workspace crons/ directory."""
+
+    name: str = Field(description="Unique job identifier")
+    schedule: str = Field(description="Cron expression (e.g., '0 9 * * *')")
+    enabled: bool = Field(default=True, description="Whether the job is active")
+    prompt: str = Field(description="User prompt to inject when cron triggers")
+    output: CronOutputConfig = Field(description="Where to send the response")
+
+    @field_validator("schedule")
+    @classmethod
+    def validate_cron_expression(cls, v: str) -> str:
+        """Validate cron expression is parseable at config load time."""
+        try:
+            CronTrigger.from_crontab(v)
+        except (ValueError, KeyError) as e:
+            raise ValueError(f"Invalid cron expression '{v}': {e}") from e
+        return v
 
 
 class WorkspaceConfig(BaseModel):
