+++ pyproject.toml
@@ -73,11 +73,6 @@ line-length = 120
 [tool.ruff.lint]
 select = ["E", "F", "I", "N", "W", "UP"]
 
-[tool.ruff.lint.per-file-ignores]
-# Deprecation shims: Allow imports after warnings
-"openpaw/commands/*.py" = ["E402"]
-"openpaw/commands/handlers/*.py" = ["E402"]
-
 [tool.mypy]
 python_version = "3.11"
 strict = true
