+++ tests/test_approval_gates.py
@@ -6,7 +6,7 @@ from unittest.mock import AsyncMock, MagicMock
 import pytest
 
 from openpaw.core.config.models import ApprovalGatesConfig, ToolApprovalConfig
-from openpaw.stores.approval import ApprovalGateManager
+from openpaw.runtime.approval import ApprovalGateManager
 from openpaw.agent.middleware.approval import ApprovalRequiredError, ApprovalToolMiddleware
 from openpaw.agent.runner import AgentRunner
